% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{astho_governance}
\alias{astho_governance}
\title{ASTHO Profile data on state and territorial health agency and jurisdiction characteristics}
\format{
\subsection{\code{astho_governance}}{

A data frame with 59 rows and 8 columns:
\describe{
\item{fips}{Federal Information Processing System (FIPS) Codes}
\item{name}{Jurisdiction Name}
\item{governance}{Agency Governance Classification}
\item{size}{Population Size}
\item{hhs_region}{Health and Human Services Regions}
\item{density}{Population Density}
\item{structure}{Agency Structure}
\item{region}{Combined HHS Regions}
}
}
}
\source{
Association of State and Territorial Health Officials (ASTHO) Profile of State and Territorial Public Health Survey, 2022.
Population size from: (1) United States Census Bureau. Population Estimates Program. About the Population Estimates Program, and (2) United Nations, Department of Economic and Social Affairs, Population Division. World Population Prospects.
Density from: (1) United States Census Bureau. Population Estimates Program. About the Population Estimates Program, (2) United States Census Bureau. Geography Reference Files. 2010 Census Geographies by Area, and (3) United Nations, Department of Economic and Social Affairs, Population Division. World Population Prospects.
}
\usage{
astho_governance
}
\description{
A subset of data from the ASTHO Profile of State and Territorial Public Health Survey, 2022.
Report ...
}
\keyword{datasets}
