% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_all}
\alias{map_all}
\title{Create an accessible categorical map using the ASTHO theme in {highcharter}}
\usage{
map_all(
  data,
  fips_column,
  selected_value,
  my_colors = main_colors,
  my_click = NA,
  title_text,
  subtitle_text,
  caption_text,
  legend_text = NA,
  accessible_desc,
  my_tooltip = "{point.value}",
  my_reverse = FALSE,
  enable_legend = TRUE
)
}
\arguments{
\item{data}{the data frame to create the map.}

\item{fips_column}{column of a dataframe with the Federal Information Processing System (FIPS) Codes for States and Counties, in quotes.}

\item{selected_value}{column of a dataframe with the values to visualize in the map, in quotes.}

\item{my_colors}{vector; colors to supply to the map, as hex values.}

\item{my_click}{JS function; use to add a click event, requires a JS function, JS(), Defaults to NA.}

\item{title_text}{character string; title of the map.}

\item{subtitle_text}{character string; subtitle to appear underneath the title, typically a date or year.}

\item{caption_text}{character string; source line and data notes to appear under the map.}

\item{legend_text}{character string; title of the legend. Defaults to NA.}

\item{accessible_desc}{character string; alternative text description of the figure for screen readers.}

\item{my_tooltip}{character string; text and values to appear in the tooltip, use {point.xx} to call a value from the data frame, where xx is a specific column.}

\item{my_reverse}{boolean, TRUE or FALSE, where TRUE reverses the legend. Defaults to FALSE.}

\item{enable_legend}{boolean, TRUE or FALSE, where TRUE enables the legend. Defaults to TRUE.}
}
\value{
A categorical map
}
\description{
Create an accessible categorical map using the ASTHO theme in {highcharter}
}
\examples{

map_all(data = astho_governance,
      fips_column = "fips",
       selected_value = "governance",
       title_text = "Map of U.S. State and Island Area Public Health Agency Governance", # title
       subtitle_text = "2022", # content under the title
       caption_text =
       "<b> Source: </b> Association of State and Territorial Health Officials.
       ASTHO Profile of State and Territorial Public Health, Volume Six. Arlington,
       VA: Association of State and Territorial Health Officials. 2023.",
      accessible_desc =
      "This is a map of governance classification, with sost jurisdictions being
      decentralized followed by centralized. Few jurisdictions are mixed. Enter the
      map to review classification at the jurisdiction level.",
      my_tooltip =
      "<b>{point.JurisdictionName}</b><br>
      <b>Agency Governance:</b>   {point.governance}<br>
      <b>Agency Structure:</b>   {point.structure}<br>
      <b>HHS Region:</b>   {point.HHSRegion}<br>
      <b>Population Size:</b>   {point.size}<br>",
     )


}
